+++ src/screens/TempJobsScreen.js
@@ -1,18 +1,7 @@
-import React, { useState } from "react";
+import React from "react";
 import styles from "./TempJobsScreen.module.css";
+import SearchBarJobs from "../components/SearchBarJobs";
 
-const dummyCategories = [
-  "Tutoring",
-  "Gardeing",
-  "Housekeeping",
-  "Programming",
-  "Accounting",
-  "Marketing",
-];
-
-const CatButton = (props) => {
-  return <button className={styles.catButton}>{props.title}</button>;
-};
 
 const TempJobsScreen = () => {
   return (
@@ -24,7 +13,7 @@ const TempJobsScreen = () => {
         <h5 style={{ marginLeft: 20, marginTop: -14, paddingRight: 20 }}>
           Search for individuals that can help you with labor
         </h5>
-
+        <SearchBarJobs />
       </div>
     </div>
   );
